Implement API versioning
Pat of https://github.com/SneaksAndData/boxer-issuer/issues/65
