+++ src/http/controllers.rs
@@ -1,0 +1,1 @@
+pub mod v1;
