--- src/http/controllers/mod.rs
@@ -1,5 +1,0 @@
-pub mod identity;
-pub mod principal;
-pub mod provider;
-pub mod schema;
-pub mod token;
