+++ src/http/controllers/v1/provider/oidc_provider_registration.rs
