+++ src/http/openapi.rs
@@ -1,31 +1,10 @@
-use crate::http::controllers;
-use utoipa::openapi::security::{Http, HttpAuthScheme, SecurityScheme};
-use utoipa::{Modify, OpenApi};
+use crate::http::controllers::v1::ApiV1;
+use utoipa::OpenApi;
 
 #[derive(OpenApi)]
-#[openapi(paths(
-    controllers::identity::post_identity,
-    controllers::identity::get_identity,
-    controllers::identity::delete_identity,
-    controllers::schema::post_schema,
-    controllers::schema::get_schema,
-    controllers::schema::delete_schema,
-    controllers::principal::post_principal,
-    controllers::principal::get_principal,
-    controllers::provider::post_provider,
-    controllers::provider::get_provider,
-    controllers::provider::delete_provider,
-    controllers::token::token,
-),
-    modifiers(&SecurityAddon))]
+#[openapi(
+    nest(
+        (path = "/api/v1", api = ApiV1)
+    ),
+)]
 pub struct ApiDoc;
-
-struct SecurityAddon;
-
-impl Modify for SecurityAddon {
-    fn modify(&self, openapi: &mut utoipa::openapi::OpenApi) {
-        let components = openapi.components.as_mut().unwrap(); // we can unwrap safely since there already is components registered.
-        components.add_security_scheme("external", SecurityScheme::Http(Http::new(HttpAuthScheme::Bearer)));
-        components.add_security_scheme("internal", SecurityScheme::Http(Http::new(HttpAuthScheme::Bearer)));
-    }
-}
