+++ src/services/backends/kubernetes/identity_repository.rs
@@ -1,7 +1,6 @@
 mod external_identity;
 
-// Workaround to use prinltn! for logs.
-use crate::http::controllers::identity::external_identity_registration::ExternalIdentityRegistration;
+use crate::http::controllers::v1::identity::external_identity_registration::ExternalIdentityRegistration;
 use crate::services::backends::kubernetes::identity_repository::external_identity::ExternalIdentityDocument;
 use boxer_core::services::backends::kubernetes::kubernetes_resource_manager::status::Status;
 use boxer_core::services::backends::kubernetes::repositories::KubernetesRepository;
